+++ commons/src/constants/attachment-content-types.ts
@@ -1,0 +1,5 @@
+export const DisallowedContentTypes = [
+  "message/delivery-status",
+  "message/rfc822",
+  "text/calendar",
+];