+++ components/email/src/Email.svelte
@@ -3,6 +3,7 @@
 <script lang="ts">
   import {
     EmailStore,
+    ErrorStore,
     ManifestStore,
     fetchAccount,
     updateThread,
@@ -48,9 +49,14 @@
   import * as DOMPurify from "dompurify";
   import LoadingIcon from "./assets/loading.svg";
   import { downloadFile } from "@commons/connections/files";
+  import { DisallowedContentTypes } from "@commons/constants/attachment-content-types";
 
   const dispatchEvent = getEventDispatcher(get_current_component());
   $: dispatchEvent("manifestLoaded", manifest);
+
+  $: if (Object.keys($ErrorStore).length) {
+    dispatchEvent("onError", $ErrorStore);
+  }
 
   export let id: string = "";
   export let access_token: string = "";
@@ -129,7 +135,7 @@
   });
 
   let previousProps = $$props;
-  $: {
+  $: (async () => {
     if (JSON.stringify(previousProps) !== JSON.stringify($$props)) {
       _this = buildInternalProps(
         $$props,
@@ -137,16 +143,20 @@
         defaultValueMap,
       ) as EmailProperties;
 
-      transformPropertyValues();
+      await transformPropertyValues();
       previousProps = $$props;
     }
-  }
+  })();
 
-  function transformPropertyValues() {
+  async function transformPropertyValues() {
     _this.thread_id = !thread && !message_id && !message ? _this.thread_id : "";
 
     if (id && !_this.thread_id && !_this.thread && _this.message_id) {
       fetchOneMessage();
+    }
+
+    if (!activeThread || previousProps.thread_id !== _this.thread_id) {
+      await getThread();
     }
   }
 
@@ -237,16 +247,40 @@
   let activeThread: Conversation;
 
   // #region thread intake and set
-  // The trick is to always ensure that activeThread is in the store; that way if we need to do fetches to update its messages, it too will be updated for free.
-  // TODO: this feels like it could be a "$: activeThread =" reactive prop declaration instead of a conditional block. -Phil
-  $: (async () => {
+  let fetchingThreadPromise: Promise<Conversation>;
+  async function getThread() {
+    if (id && _this.thread_id && !fetchingThreadPromise) {
+      fetchingThreadPromise = <Promise<Conversation>>fetchThread({
+        component_id: id,
+        thread_id: _this.thread_id,
+        access_token,
+      }).catch(silence);
+      const thread = await fetchingThreadPromise;
+      fetchingThreadPromise = <any>null;
+
+      // get body for last message in open thread
+      if (thread?.messages?.length > 0) {
+        const lastMsgIndex = thread.messages.length - 1;
+        thread.messages[lastMsgIndex].body = await fetchIndividualMessage(
+          lastMsgIndex,
+          thread.messages[lastMsgIndex].id,
+        );
+      }
+      if (thread) {
+        thread.expanded =
+          activeThread?.expanded ?? _this.show_expanded_email_view_onload;
+        activeThread = thread;
+      }
+    }
+  }
+
+  $: {
     if (_this.thread && _this.thread.id) {
       // Is it in the store already? (via <nylas-mailbox>, for example)
-      const localThread: Conversation =
-        $EmailStore[queryKey]?.find(
-          (storedThread: Conversation) =>
-            storedThread && storedThread.id === thread?.id,
-        ) ?? (_this.thread as Conversation);
+      const localThread = ($EmailStore[queryKey]?.find(
+        (storedThread: Thread) =>
+          storedThread && storedThread.id === thread?.id,
+      ) ?? _this.thread) as Conversation;
 
       // This is for Email component demo purpose, where we want to show expanded threads by default on load.
       if (_this.show_expanded_email_view_onload) {
@@ -257,27 +291,8 @@
         lastMsg.body = lastMsg.body ?? lastMsg.snippet;
       }
       activeThread = localThread;
-    } else if (_this.thread_id) {
-      const thread = await fetchThread(query).catch(silence);
-
-      if (thread) {
-        if (_this.show_expanded_email_view_onload) {
-          thread.expanded = _this.show_expanded_email_view_onload;
-
-          // get body for last message in open thread
-          if (thread.messages.length) {
-            const lastMsgIndex = thread.messages.length - 1;
-            thread.messages[lastMsgIndex].body = await fetchIndividualMessage(
-              lastMsgIndex,
-              thread.messages[lastMsgIndex].id,
-            );
-          }
-        }
-
-        activeThread = thread;
-      }
     }
-  })();
+  }
 
   // #endregion thread intake and set
   let emailManuallyPassed: boolean;
@@ -323,12 +338,7 @@
     }
   }
 
-  $: if (
-    !_this.thread &&
-    _this.clean_conversation &&
-    ((activeThread && !activeThread.messages.some((m) => m.conversation)) ||
-      (_this.message && !_this.message.conversation))
-  ) {
+  $: if (_this.clean_conversation && (_this.thread_id || _this.message_id)) {
     cleanConversation();
   }
   //#endregion Clean Conversation
@@ -397,7 +407,6 @@
       }
 
       //#region open thread + messages
-
       activeThread.expanded = !activeThread.expanded;
       // Upon expansion / lastMessage existing, scroll to it
       if (activeThread.expanded && _this.click_action === "default") {
@@ -534,15 +543,18 @@
     msgIndex: number,
     messageID: string,
   ): Promise<string | null> {
-    messageLoadStatus[msgIndex] = "loading";
-    return fetchMessage(query, messageID).then(async (json) => {
-      messageLoadStatus[msgIndex] = "loaded";
-      if (FilesStore.hasInlineFiles(json)) {
-        const messageWithInlineFiles = await getMessageWithInlineFiles(json);
-        return messageWithInlineFiles.body;
-      }
-      return json.body;
-    });
+    if (id) {
+      messageLoadStatus[msgIndex] = "loading";
+      return fetchMessage(query, messageID).then(async (json) => {
+        messageLoadStatus[msgIndex] = "loaded";
+        if (FilesStore.hasInlineFiles(json)) {
+          const messageWithInlineFiles = await getMessageWithInlineFiles(json);
+          return messageWithInlineFiles.body;
+        }
+        return json.body;
+      });
+    }
+    return new Promise(() => null);
   }
 
   function doNothing(e: MouseEvent | KeyboardEvent) {
@@ -666,13 +678,16 @@
 
   let attachedFiles: Record<string, File[]> = {};
 
-  $: activeThread, activeThread ? initializeAttachedFiles() : "";
+  $: activeThread ? initializeAttachedFiles() : "";
 
   function initializeAttachedFiles() {
-    attachedFiles = activeThread.messages.reduce(
+    attachedFiles = activeThread.messages?.reduce(
       (files: Record<string, File[]>, message) => {
         for (const [fileIndex, file] of message.files.entries()) {
-          if (file.content_disposition === "attachment") {
+          if (
+            file.content_disposition === "attachment" &&
+            !DisallowedContentTypes.includes(file.content_type)
+          ) {
             if (!files[message.id]) {
               files[message.id] = [];
             }
@@ -701,7 +716,7 @@
     return message;
   }
 
-  async function downloadSelectedFile(event: CustomEvent, file: File) {
+  async function downloadSelectedFile(event: MouseEvent, file: File) {
     event.stopImmediatePropagation();
     if (id && ((activeThread && _this.thread_id) || _this.message_id)) {
       const downloadedFileData = await downloadFile({
@@ -718,8 +733,8 @@
     });
   }
 
-  async function handleDownloadFromMessage(event: CustomEvent) {
-    const file = event.detail.file;
+  async function handleDownloadFromMessage(event: MouseEvent) {
+    const file = (<any>event.detail).file;
     downloadSelectedFile(event, file);
   }
 </script>
@@ -837,10 +852,7 @@
         }
 
         .thread-message-count {
-          color: var(
-            --nylas-email-thread-message-count-color,
-            var(--grey-light)
-          );
+          color: var(--nylas-email-thread-message-count-color, var(--black));
           font-size: 12px;
           align-self: center;
         }
@@ -924,6 +936,11 @@
               background: none;
               display: flex;
               cursor: pointer;
+
+              * {
+                width: 0.7em;
+                height: 0.7em;
+              }
             }
           }
           [role="toolbar"] {
@@ -1138,6 +1155,11 @@
               background: none;
               cursor: pointer;
               display: flex;
+
+              * {
+                width: 1em;
+                height: 1em;
+              }
             }
           }
         }
@@ -1282,6 +1304,12 @@
             display: flex;
             align-items: center;
             justify-content: center;
+
+            .spinner {
+              height: 18px;
+              animation: rotate 2s linear infinite;
+              margin-right: 10px;
+            }
           }
           @keyframes rotate {
             to {
@@ -1370,15 +1398,6 @@
                 <h1>{thread?.subject}</h1>
               </div>
               <div role="toolbar">
-                <div class="delete">
-                  <button
-                    title="Delete thread / Move to trash"
-                    aria-label="Delete thread (Move to trash)"
-                    on:click|stopPropagation={(e) => deleteEmail(e)}
-                  >
-                    <TrashIcon />
-                  </button>
-                </div>
                 {#if _this.show_star}
                   <div class="starred">
                     <button
@@ -1391,24 +1410,36 @@
                       aria-checked={thread.starred}
                       on:click|stopPropagation={handleThreadStarClick}
                     />
-                  </div>{/if}
-                <div class="read-status">
-                  <button
-                    title={`Mark thread as ${
-                      activeThread.unread ? "" : "un"
-                    }read`}
-                    aria-label={`Mark thread as ${
-                      activeThread.unread ? "" : "un"
-                    }read`}
-                    on:click|stopPropagation={toggleUnreadStatus}
-                  >
-                    {#if activeThread.unread}
-                      <MarkReadIcon aria-hidden="true" />
-                    {:else}
-                      <MarkUnreadIcon aria-hidden="true" />
-                    {/if}
-                  </button>
-                </div>
+                  </div>
+                {/if}
+                {#if _this.show_thread_actions}
+                  <div class="delete">
+                    <button
+                      title="Delete thread / Move to trash"
+                      aria-label="Delete thread (Move to trash)"
+                      on:click|stopPropagation={(e) => deleteEmail(e)}
+                    >
+                      <TrashIcon />
+                    </button>
+                  </div>
+                  <div class="read-status">
+                    <button
+                      title={`Mark thread as ${
+                        activeThread.unread ? "" : "un"
+                      }read`}
+                      aria-label={`Mark thread as ${
+                        activeThread.unread ? "" : "un"
+                      }read`}
+                      on:click|stopPropagation={toggleUnreadStatus}
+                    >
+                      {#if activeThread.unread}
+                        <MarkReadIcon aria-hidden="true" />
+                      {:else}
+                        <MarkUnreadIcon aria-hidden="true" />
+                      {/if}
+                    </button>
+                  </div>
+                {/if}
               </div>
             </header>
             {#if activeThread.messages.length}
@@ -1530,10 +1561,7 @@
                         {/await}
                       {:else}
                         <div class="email-loader">
-                          <LoadingIcon
-                            class="spinner"
-                            style="height:18px; animation: rotate 2s linear infinite; margin-right:10px;"
-                          />
+                          <LoadingIcon class="spinner" />
                           Loading...
                         </div>
                       {/if}
@@ -1658,9 +1686,9 @@
                     {/if}
                   </div>
                 </div>
-                {#if activeThread.messages.length > 1 && _this.show_number_of_messages}
-                  <span class="thread-message-count"
-                    >{activeThread.messages.length}
+                {#if _this.show_number_of_messages && activeThread?.messages?.length > 0}
+                  <span class="thread-message-count">
+                    {activeThread.messages.length}
                   </span>
                 {/if}
               </div>
@@ -1668,9 +1696,8 @@
             <div class="subject-snippet-date">
               <div class="snippet-attachment-container">
                 <div class="desktop-subject-snippet">
-                  <span class="subject">{thread?.subject}</span><span
-                    class="snippet"
-                  >
+                  <span class="subject">{thread?.subject}</span>
+                  <span class="snippet">
                     {thread.snippet}
                   </span>
                 </div>
@@ -1678,7 +1705,10 @@
                   <div class="attachment desktop">
                     {#each Object.values(attachedFiles) as files}
                       {#each files as file}
-                        <button on:click={(e) => downloadSelectedFile(e, file)}>
+                        <button
+                          on:click={(event) =>
+                            downloadSelectedFile(event, file)}
+                        >
                           {file.filename || file.id}
                         </button>
                       {/each}
@@ -1717,8 +1747,8 @@
                         <MarkReadIcon aria-hidden="true" />
                       {:else}
                         <MarkUnreadIcon aria-hidden="true" />
-                      {/if}</button
-                    >
+                      {/if}
+                    </button>
                   </div>
                 {:else if _this.show_received_timestamp}
                   <span>
@@ -1739,7 +1769,9 @@
                 <div class="attachment mobile">
                   {#each Object.values(attachedFiles) as files}
                     {#each files as file}
-                      <button on:click={(e) => downloadSelectedFile(e, file)}>
+                      <button
+                        on:click={(event) => downloadSelectedFile(event, file)}
+                      >
                         {file.filename || file.id}
                       </button>
                     {/each}
@@ -1819,11 +1851,13 @@
                 {/if}
               </div>
             </div>
-            <div class="message-date">
-              <span>
-                {formatPreviewDate(new Date(_this.message.date * 1000))}
-              </span>
-            </div>
+            {#if _this.show_received_timestamp}
+              <div class="message-date">
+                <span>
+                  {formatPreviewDate(new Date(_this.message.date * 1000))}
+                </span>
+              </div>
+            {/if}
           </div>
           <div class="message-body">
             {#if _this.clean_conversation && message.conversation}