+++ components/agenda/CHANGELOG.md
@@ -4,10 +4,17 @@
 
 ## New Features
 
+- [Agenda] Added `onError` custom event when an error occurs [#262](https://github.com/nylas/components/pull/262)
+  
+## Bug Fixes
+
+# v1.1.5 (2021-12-10)
+
 ## Bug Fixes
 
 - [Agenda] Ensure that the `auto_time_box` property updates when events update [#243](https://github.com/nylas/components/pull/243)
-# v1.1.4
+
+# v1.1.4 (2021-12-09)
 
 ## Bug Fixes