+++ components/composer/CHANGELOG.md
@@ -4,9 +4,17 @@
 
 ## New Features
 
+- [Composer] Added `onError` custom event when an error occurs [#262](https://github.com/nylas/components/pull/262)
+
 ## Bug Fixes
 
-# v1.1.4
+# v1.1.5 (2021-12-10)
+
+## Bug Fixes
+
+- Added better error handling when file size exceeds 5MB [248](https://github.com/nylas/components/pull/248)
+
+# v1.1.4 (2021-12-09)
 
 ## Bug Fixes