+++ components/email/CHANGELOG.md
@@ -3,11 +3,22 @@
 ## New Features
 
 - Participants list truncates after 1 participant and displays how many more participants are included. Clicking 'And {N} more' opens a tooltip with additional participants.
+- [Email] Added `onError` custom event when an error occurs [#262](https://github.com/nylas/components/pull/262)
+
+## Bug Fixes
+
+# v1.1.5 (2021-12-10)
+
+## New Features
+
+- Added support for downloading attachments [#240](https://github.com/nylas/components/pull/240)
+- Added support for rendering inline images (broken images) [249](https://github.com/nylas/components/pull/249)
 
 ## Bug Fixes
 - Triggering a tooltip for an individual participant no longer opens tooltip for all participants.
+- Fixed props inconsistencies [253](https://github.com/nylas/components/pull/253)
 
-# v1.1.4
+# v1.1.4 (2021-12-09)
 
 ## Bug Fixes