+++ components/mailbox/CHANGELOG.md
@@ -4,9 +4,20 @@
 
 ## New Features
 
+- [Mailbox] Added `onError` custom event when an error occurs [#262](https://github.com/nylas/components/pull/262)
+
 ## Bug Fixes
 
-# v1.1.4
+# v1.1.5 (2021-12-10)
+
+## New Features
+
+- Added support for downloading attachments [#240](https://github.com/nylas/components/issues/240)
+- Added support for rendering inline images (broken images) [249](https://github.com/nylas/components/pull/249)
+
+## Bug Fixes
+
+# v1.1.4 (2021-12-09)
 
 ## Bug Fixes